Implement is_closed / set_closed functionality for EtcdRendezvous
Summary:
`set_closed` is used when a worker decides the job as a whole is finished - it will mark the rendezvous closed and exit. Other workers will (eventually) observe the rendezvous closed, and will either not attempt to do a re-rendezvous, or will early-exit from an in-progress-rendezvous (by raising the RendezvousClosedException).

`is_closed` is used to query if the rdzv is known to have been closed.

This has semantics of eventual propagation, hence is not to be relied on for accurate synchronization.

How/why this implementation works:
1. During rendezvous process, all state transitions (say, A->B) happen using either CAS (if done by this worker) or wait-loop (if done by other worker), where the A state is always tested for being an allowed state for this A->B transition. In case A is not allowed for A->B transitionm, a retry is triggered from the very beginning (init phase).
2. The 'closed' state cannot be a source of any transiton, i.e. is not allowed to be A in A->B. This means a retry to init_phase is definitely going to happen. This includes early-exiting from any wait-loops.
3. Inside the init_phase, which check if the observed state is closed, and raise RendezvousClosedException. Because of this design where we always fallback to init_phase for interim state transitions, we only have one place where we need to handle the closed condition.

Note that even "create" and "delete" state transitions are performed using CAS, which is important for the correctness argument.

Reviewed By: kiukchung

Differential Revision: D18759684

fbshipit-source-id: 7348b901ebbe584d6afb6ab638d49d5d7b5ae6b0