Load checkpoint only when none of the trainers did
Summary:
we load checkpoint only if all trainers start from scratch. it is not necessary to load checkpoint if there is a good trainer as new trainer can sync state from it.

Today we ask trainer 0 (0 in distributed rank) to load checkpoint, however, there is a samll chance that after re-rendezvous we assigned rank 0 to a new trainer and the previous tainer0 is still alove, since the `checkpoint_loaded` is False, we will load checkpoint again.

Load checkpoint twice is not necessary, and we might also cause state de-synchronize if we didn't implement `sync` correct.

The fix here is do a collective comminucation to detemine whether we need to reload checkpoint.

Reviewed By: kiukchung

Differential Revision: D19109746

fbshipit-source-id: 04f606245ae8dd1f6eeac4e4d7a284fec2373a33