Introduce PET state generator
Summary: Today PET requires a train_step and runs a managed train_loop, PET controls the training process. This is not friendly to applications that have their own customized training loop, which make user adoption a little challenge. To mitigate this problem we propose *PET state generator*. It treat customized train_loop as a python generator that yields a sequence of PET states. This state generator approach returns the control to user code and simply the programing model very much.

Differential Revision: D19910960

fbshipit-source-id: 5941edd6e0f66fb30f58dfffa7e62fda71da0b22