introduce run_train that takes a train_step generator function (#44)
Summary:
Pull Request resolved: https://github.com/pytorch/elastic/pull/44

Today PET requires a train_step and runs a managed train_loop, PET controls the training process. This is not friendly to applications that have their own customized training loop, which make user adoption a little challenge. To mitigate this problem we propose *PET state generator*. It treat customized train_loop as a python generator that yields a sequence of PET states. This state generator approach returns the control to user code and simply the programing model very much.

Reviewed By: isunjin

Differential Revision: D19910960

fbshipit-source-id: 1bc77a10fc23da654ab4b5b214ef3de3063c01cc