expose run_id as TORCHELASTIC_RUN_ID env var for workers
Summary: run_id is a unique id for the run instance of the job. It is useful for the workers to have access to this information for various purposes. For instance it serves a a prefix for a directory that local rank 0 creates that would not collide with other jobs that are stacked on the same node.

Differential Revision: D22104173

fbshipit-source-id: 79967199591007a7316035088ebf59f0fb838e18