Introduce shutdown method for rendezvous API
Summary:
Introduce shutdown method for rendezvous API. Currently the shutdown behavior is implemented in destructor. Unfortunately, it is impossible to know when the destructor is invoked if the c++ code is integrated with python. This can lead to bugs.

One bug that we see is that the destructor can be invoked after test class method is finished, which can lead to timeout issues.

Reviewed By: kiukchung

Differential Revision: D22352117

fbshipit-source-id: 373fa2bf164644893fb0e512531e636006463d49