Introduce runopt and RunConfig for scheduler
Summary:
Part [1/n] in making unbinding a particular scheduler from session work. End
 goal is to support `session.run(app, runconfig, scheduler_backend="local ")` or `session.run(app, runconfig, scheduler=LocalScheduler())`

See pydocs for details on what `runopt` and `RunConfig` classes are. TL;DR `runopt` is similar to `argparse` (it holds the supported run options for each scheduler). `RunConfig` is analogous to the actual CMD args for a python binary.

Reviewed By: yifuwang

Differential Revision: D24472051

fbshipit-source-id: f1046933cc47a3da116718cd76283f3b631f3e04