Resolve bug when workers hang due to huge output
Summary:
By default, torch and torchelastic use the python multiprocessing module to communicate between processes. Multiprocessing.queue module uses standard linux pipes to perform IPC. If the huge object is transferred between processes, the process that writes data into the pipe can hang forever if the reader process does not start reading the pipe. By default the size of the pipe is 4096*16= 65536 bytes . Everything that is greater than that will stall the pipe.

This diff changes the way we monitor workers.

Reviewed By: kiukchung

Differential Revision: D24467961

fbshipit-source-id: 51f1cb4260bfcf5baeb99fe044769d497cd2dd91