randomize local scheduler app id
Summary:
One of the most prevalent use case of local scheduler is running a program that
launches one job. Sometime for debugging purposes this is done repeatedly and
often the program doesn't shutdown gracefully.

With incremental app id, users are reusing the same PET run id repeatedly in
this scenario. The residual of previous runs could cause new runs to hang in
rendezvous.

Reviewed By: kiukchung

Differential Revision: D24456538

fbshipit-source-id: c8a8ac18f60d19e5baba9d13cd602ef4dbda1422