Refactor error-handler and torchelastic to make it return-based when errors occur rather than exception-based
Summary:
The diff introduces the following changes:
1. Make torchelastic.multiprocessing.error_handler return failures instead of throwing exception
2. Make torchelastic overall return-based on failures rather than exception-based.
3. Remove WorkerGroupFailureException class
4. Even though the agent does not throw exceptions if worker errors occur, launchers raise exceptions to indicate to the schedulers that something went wrong.
5. Makes error handler file based instead of dir based.

The error location now controlled by TORCHELASTIC_ERROR_FILE, which is set during mp_context creation. If the parent(agent) does not call torchelastic.mp.errors.record method, the parent error file is set to : os.path.join(temp_dir, error.log), otherwise the parent error file will be set to scheduler-specific location.
The worker file always will be in the same directory as parent file, and has a template:
"os.path.join(parent_dir, rank, error.log)"

The launchers are responsible for handling failures produced by agent. The current behavior is that the launchers copy-paste the content to the scheduler specific destination and raise the exception with a fail message. If there were no files written, the record on the launcher main function will process and write the exception to the file, otherwise the exception will be only written to the stderr.

Reviewed By: kiukchung

Differential Revision: D24710390

fbshipit-source-id: 719280205b28b4c02ee7beaa61bcdcb904aaa7a8