Implement agent shutdown method
Summary:
The diff moves the actual process start logic out of the constructor into separate start method, which makes it closer to the context manager pattern. This change allows to safely close any open resources and threads if errors occur during process start.

The diff also introduces a shutdown method for agent, that is used by the agent's run method in the finally block. If agent allocates any resources, e.g. opening thread, or files, the resources had to be closed in the shutdown method.

The resources are expected to be allocated per agent run.

Reviewed By: kiukchung

Differential Revision: D25357931

fbshipit-source-id: 49105d7d6d462d5cd4282f06f45d3e9372cbd48f