Use spawn instead of fork
Summary:
The diff modifies local elastic agent tests to use spawn method instead of fork. Currently there is a caveat in using spawn: When the test that requires multiple processes and some kind of agreement gets executed via python code coverage, it gets deadlocked.

I do not know the final reason why it happens, but I assume that python code coverage does not execute the processes with spawn methods.

As a mitigation we add additional parameter to the run_job that control whether tests should use the main process or not.

Reviewed By: kiukchung

Differential Revision: D26013470

fbshipit-source-id: 4d1c0258aa5a94f8bde2af18750d629db4ce18d5