Decode new line before dumping object to the json, add run time for agent event records
Summary:
The raw_error that is accepted by the event is the encoded json. Since the event by itself is the json, there is double json encode happens. The diff adds decoding of the new line to avoid double decoding it

The diff also adds run time to the agent event records.

Reviewed By: kiukchung

Differential Revision: D26218902

fbshipit-source-id: 5e16e28d4c8382b287bfd1f9cace31b3150d3900