Modify AppStatus and DescripeAppResponse to return information about nodes
Summary:
The diff introduces a new dataclass: NodeStatus which is used to describe the state of the node. The dataclass is used when querying the app status.

The diff also adds pretty print to the app status to clearly present a picture about the node states.

Note: for large jobs it makes sense to introduce state filters

Reviewed By: kiukchung

Differential Revision: D26118306

fbshipit-source-id: 7b2478f6e83c3d1b4e41a2b74dfbb37612535dff