Introduce app metadata
Summary:
The diff introduces app metadata that is used by the mast_scheduler.
Note: The metadata is mutable and overridable dict and could be set multiple times. This is mostly done for flexibility and ability to override metadata when necessary.

The diff also moves rm_attribution inside the App metadata dict.

Reviewed By: kiukchung

Differential Revision: D26802009

fbshipit-source-id: 71ba470040ad0a2cf9eaac0631946478a5d02898